[SUPERVISOR D ZB]

+++ philo_bonus/srcs/subroutines/runProcesses.c
@@ -1,7 +1,7 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   spawnProcs.c                                       :+:      :+:    :+:   */
+/*   runProcesses.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: aabounak <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
@@ -27,17 +27,36 @@ void	*routine(void *arg)
 	return (EXIT_SUCCESS);
 }
 
-void	spawnProcs(t_status *status)
+void	*superviserRoutine(void *arg)
+{
+	t_status	*status;
+	int			count;
+	// int			i;
+
+	status = (t_status *)arg;
+	count = -1;
+	// while (++count < status->numberOfTimesMustEat)
+	// {
+	// 	i = -1;
+	// 	while (++i < status->numberOfPhilosophers)
+	// 		sem_wait(status->philos[i].eat);
+	// }
+	// sem_wait(status->write);
+	for (int j = 0; j < status->numberOfPhilosophers; j++)
+		kill(status->philos[j].pid, SIGKILL);
+	return (NULL);
+}
+
+void	runProcesses(t_status *status)
 {
 	int			i;
 	int			stat_loc;
 	pthread_t	t_id;
-	pid_t		wpid;
 
 	i = -1;
 	status->startingTime = getCurrentTime();
 	if (status->numberOfTimesMustEat > 0)
-
+		pthread_create(&t_id, NULL, &superviserRoutine, (void *)status);
 	while (++i < status->numberOfPhilosophers)
 	{
 		status->philos[i].pid = fork();
@@ -45,19 +64,19 @@ void	spawnProcs(t_status *status)
 			exit(printError("Error forking\n"));
 		if (status->philos[i].pid == 0)
 		{
-			pthread_create(&t_id, NULL, threadsSupervisor, &status->philos[i]);
+			pthread_create(&t_id, NULL, &threadsSupervisor, &status->philos[i]);
 			routine(&status->philos[i]);
-			exit(EXIT_SUCCESS);
+			exit(1);
 		}
 		ft_usleep(100);
 	}
-	while ((wpid = wait(&stat_loc)) > 0)
+	while (waitpid(-1, &stat_loc, 0) > 0)
 	{
-		// if (status->philos[i].pid)
-
-		/* for (int j = 0; j < status->numberOfPhilosophers; j++)
-			kill(status->philos[j].pid, SIGKILL);
-		return ; */
+		if (WIFEXITED(stat_loc) && WEXITSTATUS(stat_loc))
+		{
+			for (int j = 0; j < status->numberOfPhilosophers; j++)
+				kill(status->philos[j].pid, SIGKILL);
+		}
 	}
 	return ;
 }
